4505: Get data for GraphQl API from interim DB tables

+++ src/main/java/io/kontur/insightsapi/model/Axis.java
@@ -13,7 +13,7 @@ public class Axis {
 
     private List<Step> steps;
 
-    private Float quality;
+    private Double quality;
 
     private List<String> quotient;
 }
